Moved SlashCommand utilities
