+++ src/modules/discord/client/discord-client.service.ts
@@ -13,22 +13,21 @@ import {
 import { DiscordConfig } from "../../../modules/config/config";
 import { In, Repository } from "typeorm";
 import { GuildSettings } from "../models/settings.entity";
-import { getCommandMetadata, SlashCommand } from "./commands/slash-command";
+import { getCommandMetadata, SlashCommand } from "../slash-commands/slash-command";
 import { getEvents, handleEvent, On } from "./on-event";
 import { CommandBus, QueryBus } from "@nestjs/cqrs";
 import { ChannelInfo, CommunityPost } from "yt-scraping-utilities";
 import { DiscordUtil } from "../util";
-import { handleAutocomplete } from "./commands/autocomplete";
+import { handleAutocomplete } from "../slash-commands/autocomplete";
 import { Platform, SUPPORTED_PLATFORMS } from "../../../constants";
 import { Util } from "../../../util";
 import { MultipageMessage } from "../../../shared/util/multipage-message";
 import { getActions } from "../actions/decorators/action";
 import { Interval } from "@nestjs/schedule";
 import { InjectRepository } from "@nestjs/typeorm";
-import { InjectCommands } from "./commands/slash-commands.provider";
+import { InjectCommands } from "../slash-commands/slash-commands.provider";
 import { EventEmitter2, OnEvent } from "@nestjs/event-emitter";
 import { FindChannelQuery } from "../../../modules/platforms/queries";
-import { ChannelEntity } from "../../../modules/platforms/models/channel.entity";
 import { FetchPostCommand } from "../../youtube/community-posts/commands/fetch-post.command";
 
 @Injectable()
