+++ src/modules/discord/discord.module.ts
@@ -6,8 +6,8 @@ import { ActionOrchestrator } from "./actions/action.orchestrator";
 import { actionTypeFactory } from "./actions/actions-helper";
 import { StreamDiscordChannelMap } from "./actions/model/stream-thread-map.entity";
 import { ActionTypes } from "./actions/types";
-import { getCommands } from "./client/commands/slash-command";
-import { slashcommandFactory } from "./client/commands/slash-commands.provider";
+import { getCommands } from "./slash-commands/slash-command";
+import { slashcommandFactory } from "./slash-commands/slash-commands.provider";
 import { DiscordClientService } from "./client/discord-client.service";
 import { DiscordRESTService } from "./discord-rest.service";
 import { DiscordService } from "./discord.service";
