+++ src/modules/discord/slash-commands/slash-commands.provider.ts
