+++ src/modules/discord/slash-commands/types/action.command.ts
@@ -14,8 +14,8 @@ import { Repository } from "typeorm";
 import { ActionDescriptor } from "../../models/action.entity";
 import { PLATFORM_NAME_LOOKUP, Platform } from "../../../../constants";
 import { DiscordUtil } from "../../util";
-import { Autocomplete, AutocompleteReturn } from "./autocomplete";
-import { ISlashCommand, SlashCommand } from "./slash-command";
+import { Autocomplete, AutocompleteReturn } from "../autocomplete";
+import { ISlashCommand, SlashCommand } from "../slash-command";
 import { Util } from "../../../../util";
 import { InjectRepository } from "@nestjs/typeorm";
 import { EnsureChannelCommand } from "../../../platforms/commands/ensure-channel.command";
