+++ src/modules/discord/slash-commands/types/help.command.ts
@@ -1,5 +1,5 @@
 import { ChatInputCommandInteraction, CacheType, SlashCommandBuilder } from "discord.js";
-import { ISlashCommand, SlashCommand } from "./slash-command";
+import { ISlashCommand, SlashCommand } from "../slash-command";
 
 @SlashCommand({
     commandData: new SlashCommandBuilder()
